[#15]rename: 파일명 및 모듈명 변경
